Move Note and NotesList to component folder
