+++ src/App.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { BrowserRouter as Router, Route, Switch } from 'react-router-dom'
 import GlobalStyle from './GlobalStyle'
-import NotesList from './NotesList'
+import NotesList from './components/NotesList'
 import { getNotes } from './services'
 
 function App() {
