+++ src/components/Note.js
